+++ lib/neo4j/active_rel/related_node.rb
@@ -12,8 +12,6 @@ module Neo4j::ActiveRel
     #
     # Initialization with an integer happens when a relationship is loaded from the database. It loads using the ID
     # because that is provided by the Cypher response and does not require an extra query.
-    #
-    # Initialization with a node doesn't appear to happen in the code. TODO: maybe find out why this is an option.
     def initialize(node = nil)
       @node = valid_node_param?(node) ? node : (fail InvalidParameterError, 'RelatedNode must be initialized with either a node ID or node')
     end